+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 <!-- https://medium.com/@dtgasparyan/feature-sliced-design-the-ideal-frontend-architecture-84d701ad44ba -->
 ---
 ### 08 JUL 2025
+- The `useAxiosPrivate` custom hook now accepts two parameters, `accessToken` and `setAccessToken`.
 - Removed `AxiosPrivateContext` and `AxiosPrivateProvider` in `useAxiosPrivate` module.
 ---
 ### 07 JUL 2025